socket

+++ lib/component/stock/cubit/stock_cubit.dart
@@ -1,17 +1,19 @@
-import 'package:bloc/bloc.dart';
-import 'package:meta/meta.dart';
-// import 'package:socket_io_client/socket_io_client.dart';
+import 'dart:convert';
 
+import 'package:bloc/bloc.dart';
+import 'package:web_socket_channel/web_socket_channel.dart';
 part 'stock_state.dart';
 
 class StockCubit extends Cubit<StockState> {
   StockCubit() : super(StockState(counterValue: 0));
 
+  
+
   void increment() => emit(StockState(
-      counterValue: (state.counterValue as int) + 1, wasIncremented: true));
+      counterValue: (state.counterValue as double) + 1, wasIncremented: true));
 
   void decrement() => emit(StockState(
-      counterValue: (state.counterValue as int) - 1, wasIncremented: false));
+      counterValue: (state.counterValue as double) - 1, wasIncremented: false));
 
-  void replace(int value) => emit(StockState(counterValue: value));
+  void replace(double value) => emit(StockState(counterValue: value));
 }
